+++ project/src/components/shop/cart-item.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CartItem } from "@/lib/stripe";
+import { CartItem } from "@/lib/stripeUtils";
 import { useCart } from "@/contexts/CartContext";
 import { useToast } from "@/contexts/ToastContext";
 import Image from "next/image";
